Adding a more complete test with gossip-propagation of the configuration. This is still not optimal, as the Loopix should setup using its own private key, and then only exchange the public keys.
