+++ flmodules/Cargo.toml
@@ -44,7 +44,7 @@ reqwest = { version = "0.12", features = [
     "rustls-tls",
 ], default-features = false }
 bitflags = { version = "2", features = ["serde"] }
-x25519-dalek = "2"
+x25519-dalek = {version = "2", features = ["serde"]}
 hex = "0.4.3"
 bincode = "1.3.3"
 sphinx-packet = { version = "0.2" }
@@ -53,3 +53,4 @@ rand_distr = "0.4.3"
 
 [dev-dependencies]
 env_logger = "0.11"
+flmodules = {path=".", features = ["testing"]}
