+++ flmodules/src/loopix/mod.rs
@@ -1,13 +1,15 @@
-pub mod core;
-pub mod storage;
 pub mod config;
+pub mod core;
 pub mod sphinx;
+pub mod storage;
 
-pub mod provider;
 pub mod client;
 pub mod mixnode;
+pub mod provider;
 
 pub mod messages;
 
 pub mod broker;
 
+#[cfg(feature = "testing")]
+pub mod testing;
